+++ Coeus/ActorRule.h
@@ -6,10 +6,10 @@ namespace Coeus
 	class __declspec(dllexport) ActorRule : public IUpdateRule
 	{
 	public:
-		ActorRule(NetworkGradient* p_network_gradient, IUpdateRule* p_rule, double p_alpha);
+		ActorRule(NetworkGradient* p_network_gradient, IUpdateRule* p_rule, float p_alpha);
 		~ActorRule();
 
-		void calc_update(map<string, Tensor>* p_gradient, double p_delta, Tensor* p_policy, double p_alpha = 0);
+		void calc_update(map<string, Tensor>* p_gradient, float p_delta, Tensor* p_policy, float p_alpha = 0);
 		IUpdateRule* clone(NetworkGradient* p_network_gradient) override;
 		void reset() override;
 	private:
