+++ Coeus/AdaMax.h
@@ -8,7 +8,7 @@ namespace Coeus {
 		explicit AdaMax(NeuralNetwork* p_network);
 		~AdaMax();
 
-		void init(ICostFunction* p_cost_function, double p_alpha, double p_beta1 = 0.9, double p_beta2 = 0.999, double p_epsilon = 1e-8);
+		void init(ICostFunction* p_cost_function, float p_alpha, float p_beta1 = 0.9, float p_beta2 = 0.999, float p_epsilon = 1e-8);
 	};
 }
