+++ Coeus/AdaMaxRule.cpp
@@ -2,7 +2,7 @@
 
 using namespace Coeus;
 
-AdaMaxRule::AdaMaxRule(NetworkGradient* p_network_gradient, const double p_alpha, const double p_beta1, const double p_beta2, const double p_epsilon): 
+AdaMaxRule::AdaMaxRule(NetworkGradient* p_network_gradient, const float p_alpha, const float p_beta1, const float p_beta2, const float p_epsilon): 
 			IUpdateRule(p_network_gradient, p_alpha), _beta1(p_beta1), _beta2(p_beta2), _epsilon(p_epsilon)
 {
 	_m = p_network_gradient->get_empty_params();
@@ -45,7 +45,7 @@ void AdaMaxRule::reset()
 	}
 }
 
-void AdaMaxRule::calc_update(map<string, Tensor>* p_gradient, const double p_alpha) {
+void AdaMaxRule::calc_update(map<string, Tensor>* p_gradient, const float p_alpha) {
 	IUpdateRule::calc_update(p_gradient, p_alpha);
 	for (auto it = p_gradient->begin(); it != p_gradient->end(); ++it) {
