+++ Coeus/BackPropRule.cpp
@@ -2,7 +2,7 @@
 
 using namespace Coeus;
 
-BackPropRule::BackPropRule(NetworkGradient* p_network_gradient, const double p_alpha, const double p_momentum, const bool p_nesterov): IUpdateRule(p_network_gradient, p_alpha),
+BackPropRule::BackPropRule(NetworkGradient* p_network_gradient, const float p_alpha, const float p_momentum, const bool p_nesterov): IUpdateRule(p_network_gradient, p_alpha),
 	_momentum(p_momentum), 
 	_nesterov(p_nesterov)
 {
@@ -23,7 +23,7 @@ void BackPropRule::reset()
 	}
 }
 
-void BackPropRule::calc_update(map<string, Tensor>* p_gradient, const double p_alpha)
+void BackPropRule::calc_update(map<string, Tensor>* p_gradient, const float p_alpha)
 {
 	IUpdateRule::calc_update(p_gradient, p_alpha);
