+++ Coeus/CrossEntropyCost.cpp
@@ -12,13 +12,13 @@ CrossEntropyCost::~CrossEntropyCost()
 {
 }
 
-double CrossEntropyCost::cost(Tensor * p_prediction, Tensor * p_target)
+float CrossEntropyCost::cost(Tensor * p_prediction, Tensor * p_target)
 {
-	double r = 0;
+	float r = 0;
 
 	for (int i = 0; i < p_prediction->size(); i++) {
-		double e = p_target->at(i);
-		double a = p_prediction->at(i);
+		float e = p_target->at(i);
+		float a = p_prediction->at(i);
 
 		r += e * log(a) + (1 - e) * log(1 - a);
 	}
@@ -28,12 +28,12 @@ double CrossEntropyCost::cost(Tensor * p_prediction, Tensor * p_target)
 
 Tensor CrossEntropyCost::cost_deriv(Tensor * p_prediction, Tensor * p_target)
 {
-	double* data = Tensor::alloc_arr(p_prediction->size());
+	float* data = Tensor::alloc_arr(p_prediction->size());
 	int* shape = Tensor::copy_shape(p_prediction->rank(), p_prediction->shape());
 
 	for (int i = 0; i < p_prediction->size(); i++) {
-		double e = p_target->at(i);
-		double a = p_prediction->at(i);
+		float e = p_target->at(i);
+		float a = p_prediction->at(i);
 
 		data[i] = (a - e) / ((1 - a) * a);
 	}
